refactor(solver): Add branch and icp_stat
